@@ -111,14 +111,15 @@ describe('ResetPasswordPage', () => {
     );
   });
 
-  it('should reset form on success', () => {
+  it('should reset form on success', async (done) => {
     const spy = spyOn(component.formComponent.form, 'reset');
     component.success().then(() => {
       expect(spy).toHaveBeenCalledTimes(1);
     });
+    done();
   });
 
-  it('should call navigateBack with ["/users/login"] and { replaceUrl: true }, on navController when from success', () => {
+  it('should call navigateBack with ["/users/login"] and { replaceUrl: true }, on navController when from success', async (done) => {
     component.success().then(() => {
       expect(navControllerSpy.navigateBack).toHaveBeenCalledTimes(1);
       expect(navControllerSpy.navigateBack).toHaveBeenCalledWith(
@@ -126,5 +127,6 @@ describe('ResetPasswordPage', () => {
         { replaceUrl: true }
       );
     });
+    done();
   });
 });
